added Create and delete functionality to this app bruh

+++ app/src/main/java/com/example/mykip/data/User.kt
@@ -9,7 +9,7 @@ data class User(
     val nim: String,
     val email: String,
     val password: String,
-    val balance: Float = 0.00f,
+    val balance: Int = 0,
     val isAdmin: Boolean = false,
 )
